+++ src/platformAccessory.ts
@@ -1,14 +1,15 @@
-import { Service, PlatformAccessory, WithUUID, Characteristic, CharacteristicValue, CharacteristicSetCallback } from 'homebridge';
+import { Service, PlatformAccessory, WithUUID, Characteristic, CharacteristicValue, CharacteristicSetCallback, Logger } from 'homebridge';
 
 import { Zigbee2mqttPlatform } from './platform';
 import { Zigbee2mqttDeviceInfo } from './models';
 import { ExtendedTimer } from './timer';
+import { CustomServices, CustomCharacteristics } from './customServices';
 
 import * as color_convert from 'color-convert';
 
 export class Zigbee2mqttAccessory {
   public static readonly IGNORED_STATE_KEYS: Set<string> = new Set<string>(
-    ['last_seen', 'linkquality', 'voltage', 'pressure', 'illuminance', 'update_available']);
+    ['last_seen', 'linkquality', 'voltage', 'smoke_density', 'illuminance', 'update_available']);
 
   private readonly services: ServiceWrapper[] = [];
   private readonly updateTimer: ExtendedTimer;
@@ -16,15 +17,19 @@ export class Zigbee2mqttAccessory {
   private pendingPublishData: Record<string, unknown>;
   private publishIsScheduled: boolean;
 
+  get log(): Logger {
+    return this.platform.log;
+  }
+
   constructor(
     private readonly platform: Zigbee2mqttPlatform,
     public readonly accessory: PlatformAccessory,
   ) {
-    this.updateDeviceInformation(accessory.context.device);
-
     // Setup delayed publishing
     this.pendingPublishData = {};
     this.publishIsScheduled = false;
+
+    this.updateDeviceInformation(accessory.context.device);
 
     // Recreate ServiceWrappers from restored services
     for (const srv of this.accessory.services) {
@@ -67,6 +72,9 @@ export class Zigbee2mqttAccessory {
           break;
         case this.platform.Service.BatteryService.UUID:
           this.createServiceForKey('battery');
+          break;
+        case CustomServices.AirPressureSensor.UUID:
+          this.createServiceForKey('pressure');
           break;
         default:
           //ignore this service.
@@ -164,7 +172,7 @@ export class Zigbee2mqttAccessory {
 
     const remainingKeys = [...unhandledKeys].filter(k => !handledKeys.has(k));
     if (remainingKeys.length > 0) {
-      this.platform.log.debug('Unhandled keys for accessory '.concat(this.accessory.context.device.friendly_name, ': ',
+      this.log.debug('Unhandled keys for accessory '.concat(this.accessory.context.device.friendly_name, ': ',
         remainingKeys.join(', ')));
     }
   }
@@ -248,6 +256,14 @@ export class Zigbee2mqttAccessory {
         this.addService(wrapper, state, handledKeys);
         break;
       }
+      case 'pressure':
+      {
+        const wrapper = new SingleReadOnlyValueServiceWrapper('pressure',
+          this.getOrAddService(CustomServices.AirPressureSensor),
+          CustomCharacteristics.AirPressure);
+        this.addService(wrapper, state, handledKeys);
+        break;
+      }
       case 'state':
       case 'state_left':
       case 'state_right':
@@ -285,7 +301,7 @@ export class Zigbee2mqttAccessory {
   }
 
   private addService(srv: ServiceWrapper, state: Map<string, CharacteristicValue> | undefined, handledKeys: Set<string> | undefined) {
-    this.platform.log.debug(`Adding ${srv.displayName} to accessory ${this.accessory.displayName}.`);
+    this.log.debug(`Adding ${srv.displayName} to accessory ${this.accessory.displayName}.`);
     this.services.push(srv);
     let used_keys: Set<string> = new Set<string>();
     if (state !== undefined) {
@@ -306,7 +322,7 @@ export class Zigbee2mqttAccessory {
       const srv = this.services[i];
 
       if (srv.appliesToKey(key)) {
-        this.platform.log.debug(`Remove service ${srv.displayName} from ${this.accessory.displayName} (because it uses key '${key}').`);
+        this.log.debug(`Remove service ${srv.displayName} from ${this.accessory.displayName} (because it uses key '${key}').`);
         srv.remove(this.accessory);
         this.services.splice(i, 1);
       }